# pkg/agentos/openapi.yaml
openapi: 3.0.3
info:
  title: AgentOS API
  description: |
    AgentOS is a production-ready HTTP server for managing and running AI agents.

    ## Features
    - Session management for multi-turn conversations
    - Agent registration and execution
    - Support for tool-enabled agents
    - Memory and context management

    ## Authentication
    Currently no authentication is required. Authentication can be added via middleware.

  version: 1.0.0
  contact:
    name: Agno-Go Team
    url: https://github.com/jholhewres/agent-go
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: http://localhost:8080/api/v1
    description: API v1 endpoint
...
paths:
...
  /api/v1/agents/{id}/run/stream:
    post:
      tags:
        - Agents
      summary: Stream agent run events (SSE)
      description: |
        Streams agent execution events using Server-Sent Events (SSE). Clients can optionally
        filter specific event categories (e.g. `token,reasoning,complete`) via the `types` query parameter.
      operationId: streamAgentRun
      parameters:
        - name: id
          in: path
          required: true
          description: Agent ID
          schema:
            type: string
        - name: types
          in: query
          description: Comma separated list of event types to include (e.g. `run_start,token,reasoning,complete`)
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRunRequest'
      responses:
        '200':
          description: Server-sent events stream
          content:
            text/event-stream:
              schema:
                type: string
                description: SSE payload (repeating `event:` and `data:` lines)
                example: |
                  event: run_start
                  data: {"type":"run_start","data":{"input":"Hello"}}

                  event: reasoning
                  data: {"type":"reasoning","data":{"content":"thinking...","token_count":42}}

                  event: complete
                  data: {"type":"complete","data":{"output":"Hi!", "token_count":123}}
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Agent execution failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    AgentRunRequest:
      type: object
      required:
        - input
      properties:
        input:
          type: string
          description: Input text to send to the agent
          example: What is the weather in San Francisco?
        session_id:
          type: string
          description: Optional session ID for multi-turn conversations
          example: 550e8400-e29b-41d4-a716-446655440000
        stream:
          type: boolean
          description: Deprecated. Use `/api/v1/agents/{id}/run/stream` for SSE streaming.
          default: false
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: agent not found
        message:
          type: string
          description: Detailed error message
          example: agent with ID 'non-existent-agent' not found
        code:
          type: string
          description: Error code
          example: AGENT_NOT_FOUND
...